API

Modules

HeterogeneousComputing.HeterogeneousComputing

Types and constants

HeterogeneousComputing.AbstractComputeAccelerator

HeterogeneousComputing.AbstractComputeUnit

HeterogeneousComputing.AbstractGPUnit

HeterogeneousComputing.CPUnit

HeterogeneousComputing.ComputeUnitIndependent

HeterogeneousComputing.GenContext

HeterogeneousComputing.MixedComputeSystem

HeterogeneousComputing.UnknownComputeUnitOf

Functions and macros

HeterogeneousComputing.allocate_array

HeterogeneousComputing.allocate_array

HeterogeneousComputing.get_compute_unit

HeterogeneousComputing.get_compute_unit_impl

HeterogeneousComputing.get_free_memory

HeterogeneousComputing.get_total_memory

HeterogeneousComputing.ka_backend

HeterogeneousComputing.merge_compute_units

Documentation

HeterogeneousComputing.HeterogeneousComputing — Module

HeterogeneousComputing

Tools for heterogeneous computing in Julia.

HeterogeneousComputing.AbstractComputeAccelerator — Type

abstract type AbstractComputeAccelerator <: AbstractComputeUnit

Supertype for GPU compute units.

HeterogeneousComputing.AbstractComputeUnit — Type

abstract type AbstractComputeUnit

Supertype for arbitrary compute units (CPU, GPU, etc.).

adapt(cunit::AbstractComputeUnit, x) adapts x for cunit.

get_total_memory(cunit) and get_free_memory(cunit) return the total resp. the free memory on the compute unit.

allocate_array(cunit, dims) can be used to allocate new arrays on cunit.

KernelAbstractions.Backend(cunit) will return default KernelAbstractions backend for the type of the compute unit.

See also GenContext.

HeterogeneousComputing.AbstractGPUnit — Type

abstract type AbstractGPUnit <: AbstractComputeAccelerator

Supertype for GPU comute units.

HeterogeneousComputing.CPUnit — Type

struct CPUnit <: AbstractComputeUnit

CPUnit() is the default central processing unit (CPU).

HeterogeneousComputing.ComputeUnitIndependent — Type

struct ComputeUnitIndependent

get_compute_unit(x) === ComputeUnitIndependent() indicates that x is not tied to a specific compute unit. This typically means that x is a statically allocated object.

HeterogeneousComputing.GenContext — Type

GenContext{T=AbstractFloat}(
    rng::AbstractRNG = Random.default_rng(),
    cunit::AbstractComputeUnit = CPUnit()
) <: AbstractRNG

Context for generative computations.

Base.eltype(ctx::GenContext)will returnT`.

HeterogeneousComputing.MixedComputeSystem — Type

struct MixedComputeSystem <: AbstractComputeUnit

A (possibly heterogenous) system of multiple compute units.

HeterogeneousComputing.UnknownComputeUnitOf — Type

UnknownComputeUnitOf(x)

get_compute_unit(x) === ComputeUnitIndependent() indicates that the compute unit for x cannot be determined.

HeterogeneousComputing.allocate_array — Function

allocate_array(cpunit::AbstractComputeUnit, ::Type{T}, dims::Dims)
allocate_array(cpunit::AbstractComputeUnit, ::Type{T}, dims::Integer...)

Allocate a new array with element type T and size dims on compute unit cunit.

The content of the newly allocated array is undefined.

HeterogeneousComputing.allocate_array — Method

allocate_array(ctx::GenContext, dims::Dims)
allocate_array(ctx::GenContext, dims::Integer...)
allocate_array(ctx::GenContext, ::Type{T}, dims::Dims)
allocate_array(ctx::GenContext, ::Type{T}, dims::Integer...)

Allocate a new array on the compute unit and with the numerical element type specified by ctx.

The default element type can be overriden by specifying T.

HeterogeneousComputing.get_compute_unit — Function

get_compute_unit(x)::Union{
    AbstractComputeUnit,
    ComputeUnitIndependent,
    UnknownComputeUnitOf
}

Get the compute unit backing object x.

Don't specialize get_compute_unit, specialize HeterogeneousComputing.get_compute_unit_impl instead.

HeterogeneousComputing.get_compute_unit_impl — Function

HeterogeneousComputing.get_compute_unit_impl(::Type{TypeHistory}, x)::AbstractComputeUnit

See get_compute_unit_impl.

Specializations that directly resolve the compute unit based on x can ignore TypeHistory:

HeterogeneousComputing.get_compute_unit_impl(@nospecialize(TypeHistory::Type), x::SomeType) = ...

HeterogeneousComputing.get_free_memory — Function

getfreememory(cunit::AbstractComputeUnit)

Get the amount of free memory available on cunit.

HeterogeneousComputing.get_total_memory — Function

get_total_memory(cunit::AbstractComputeUnit)

Get the total amount of memory available on cunit.

HeterogeneousComputing.ka_backend — Function

HeterogeneousComputing.ka_backend(cunit::AbstractComputeUnit)

Returns the KernelAbstractions backend for cunit.

Requires KernelAbstractions.jl to be loaded, otherwise ka_backend will have no methods.

Do not call directly, use for specialization only.

User code should call KernelAbstractions.Backend(cunit) or convert(KernelAbstractions.Backend, cunit) instead, both of which will use ka_backend internally.

HeterogeneousComputing.merge_compute_units — Function

merge_compute_units(compute_units...)

Merge compute_units unto a common/combined compute unit.

Do not specialize merge_compute_units directly, specialize compute_unit_mergerule(a, b) instead.